pyproject.toml
[project]
name = "universal-ai-mcp"
version = "1.0.0"
description = "Universal AI MCP Server — modular, provider-agnostic AI assistant with planning gate and context optimization"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.9.0",
    "anthropic>=0.40.0",
    "openai>=1.58.0",
    "ollama>=0.4.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0.2",
    "tiktoken>=0.8.0",
    "aiofiles>=24.1.0",
    "starlette>=0.41.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "chromadb>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
universal-ai-mcp = "universal_ai_mcp.core.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/universal_ai_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "ANN", "S", "B", "A", "C4", "RUF"]
ignore = ["ANN101", "ANN102", "S101"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.2",
]
